+++ tests/Integration/Dto/Catalog/ProductTest.php
@@ -29,9 +29,9 @@ use Shopgate\ConnectSdk\Dto\Catalog\Product\Update;
 use Shopgate\ConnectSdk\Exception\Exception;
 use Shopgate\ConnectSdk\Exception\NotFoundException;
 use Shopgate\ConnectSdk\Exception\RequestException;
-use Shopgate\ConnectSdk\Tests\Integration\AbstractCatalogTest;
+use Shopgate\ConnectSdk\Tests\Integration\CatalogUtility;
 
-class ProductTest extends AbstractCatalogTest
+class ProductTest extends CatalogUtility
 {
     /**
      * @throws Exception