+++ tests/Integration/WebhookUtility.php
@@ -25,7 +25,7 @@ namespace Shopgate\ConnectSdk\Tests\Integration;
 use Shopgate\ConnectSdk\Dto\Webhook\Webhook;
 use Shopgate\ConnectSdk\Exception\InvalidDataTypeException;
 
-class WebhookUtility extends ShopgateSdkTest
+abstract class WebhookUtility extends ShopgateSdkUtility
 {
     const WEBHOOK_SIMPLE_PROPS = ['name', 'endpoint', 'active'];