+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- data types for fulfillment package and fulfillment package line items  
+
+### Added
+- integration tests for fulfillment package
 
 ## [1.1.1] - 2019-10-18
 ### Added